Move files in the airline GWT original project to the "airlinegwt" package so that we can have unique package names in the original projects that will not confuse the aggregate javadoc build.
