+++ gwt-parent/gwt-originals-parent/airline-gwt/src/main/java/edu/pdx/cs410J/airlinegwt/client/Flight.java
@@ -1,4 +1,4 @@
-package edu.pdx.cs410J.original.client;
+package edu.pdx.cs410J.airlinegwt.client;
 
 import edu.pdx.cs410J.AbstractFlight;
