+++ gwt-parent/gwt-originals-parent/airline-gwt/src/main/resources/edu/pdx/cs410J/airlinegwt/AirlineGwt.gwt.xml
@@ -2,7 +2,7 @@
   <inherits name="com.google.gwt.user.User"/>
   <inherits name='com.google.gwt.user.theme.standard.Standard'/>
   <inherits name='edu.pdx.cs410J.Projects'/>
-  <entry-point class="edu.pdx.cs410J.original.client.AirlineGwt"/>
+  <entry-point class="edu.pdx.cs410J.airlinegwt.client.AirlineGwt"/>
 
-  <servlet path="/ping" class="edu.pdx.cs410J.original.server.PingServiceImpl"/>
+  <servlet path="/ping" class="edu.pdx.cs410J.airlinegwt.server.PingServiceImpl"/>
 </module>
